// frontend/src/app/administrador/models/avisomes.ts
export interface Avisomes {
  id: string;
  mensajeMes: string;
  avisoLiquidoFrenosMes: number;
  avisoOperacionesSistematicasMes: number;
  avisoFiltroAireMes: number;
  avisoFiltroAireHabitaculoMes: number;
  avisoFiltroCombustibleMes: number;
  avisoFiltroAntipolenMes: number;
  avisoCorreaDistribucionMes: number;
  avisoKitDistribucionMes: number;
  avisoReglajeProyectoresMes: number;
  avisoPhLiquidoRefrigeracionMes: number;
  avisoLiquidoRefrigeracionMes: number;
  avisoCorreaArrastreAccesoriosMes: number;
  avisoKitCorreaArrastreAccesoriosMes: number;
  avisoAnticongelanteMes: number;
  avisoAceiteTransimisionMes: number;
  avisoBujiasEncendidoMes: number;
}